fix: fixed movie name rendering bug
the issue was when a movie is fetched, it is a key value of 'title' while a show is 'original_name'.

I created rendering logic hat was quite complex and confusing to show either 'movie' or 'show' depending on how it is saved in the media array.

I fixed the bug by defining 'original_name' as title when the data is fetched. This way I could render only 'title', which helped simplify the code and improve performance.